feat: update color theme
Update the color theme to use CSS variables and add a dark mode.

This change updates the color theme to use CSS variables instead of hard-coded values. This allows us to more easily change the color scheme of the site in the future.

In addition, this change adds a dark mode to the site. This can be activated by clicking the "Dark mode" button in the navigation bar.

Signed-off-by: MarJose123 <18107626+MarJose123@users.noreply.github.com>