feat(content): move payload.json to src/pages directory
This commit moves the `payload.json` file from the `public` directory to the `src/pages` directory.

This change was made to improve the organization of the project and to make it easier to manage the content of the site. The `payload.json` file contains the data that is displayed on the homepage of the site, and by moving it to the `src/pages` directory, it is now more closely associated with the code that generates the homepage.

Signed-off-by: MarJose123 <18107626+MarJose123@users.noreply.github.com>