feat: make timeline scrollable
This commit makes the timeline scrollable, allowing users to view all of their experience items without having to scroll down the page.

This is a breaking change, as it removes the `hideControls` prop from the `Chrono` component. This prop was previously used to hide the timeline controls, but it is no longer necessary with the new scrollable timeline.

Signed-off-by: MarJose123 <18107626+MarJose123@users.noreply.github.com>